Identify Login Error + Fix + Tests
