+++ package.json
@@ -5,7 +5,17 @@
     "android": "expo start --android",
     "ios": "expo start --ios",
     "web": "expo start --web",
+    "test": "jest",
     "eject": "expo eject"
+  },
+  "jest": {
+    "preset": "jest-expo",
+    "setupFilesAfterEnv": [
+      "@testing-library/jest-native/extend-expect"
+    ],
+    "transformIgnorePatterns": [
+      "node_modules/(?!(react-native|@react-native|expo|@expo|@unimodules)/)"
+    ]
   },
   "dependencies": {
     "@expo-google-fonts/nunito": "^0.2.2",
@@ -14,6 +24,7 @@
     "@react-native-community/datetimepicker": "6.5.2",
     "@react-native-masked-view/masked-view": "0.2.8",
     "@react-native-picker/picker": "^2.8.1",
+    "@react-native/assets-registry": "^0.76.1",
     "@react-navigation/native": "^6.0.16",
     "@react-navigation/stack": "^6.3.7",
     "axios": "^1.2.1",
@@ -47,10 +58,16 @@
     "uuid": "^9.0.0"
   },
   "devDependencies": {
+    "@testing-library/jest-native": "^5.0.0",
+    "@testing-library/react-native": "^11.0.0",
+    "@types/jest": "^29.5.14",
     "@types/react": "~18.0.24",
     "@types/react-dom": "~18.0.8",
     "@types/react-native": "~0.70.6",
     "@types/uuid": "^9.0.0",
+    "babel-jest": "^29.7.0",
+    "jest": "^29.0.0",
+    "jest-expo": "^51.0.4",
     "typescript": "^4.6.3"
   },
   "private": true
