+++ src/utils/index.ts
@@ -41,14 +41,6 @@ export const sanitizeEmail = (email: string): string => {
     return email.trim().toLowerCase();
 };
 
-export const validateEmail = (email: string): boolean => {
-    if (!email) return false;
-    const regex = /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{3})+$/;
-    const sanitizedEmail = email.trim().toLowerCase();
-    const result = sanitizedEmail.match(regex);
-    return !!result?.[0];
-};
-
 export const parseDateFieldFromJSONResponse = (array: [], fieldName: string): any[] => {
     return array.map((x: any) => {
         x[fieldName] = new Date(x[fieldName]);
